+++ src/components/organization/OrganizationPartner.js
@@ -1,19 +1,19 @@
 import {Box, Card, CardContent, Divider, Typography} from "@material-ui/core";
 import {colors} from "../../utils/colors";
 
-export const OrganizationPartner = ({firstname, lastname, mail}) => {
+export const OrganizationPartner = ({firstName, lastName, mail}) => {
     return (
         <Box m={2}>
             <Card style={{backgroundColor: colors.palette[8]}}>
                 <CardContent>
                     <Box m={1}>
                         <Typography>
-                            {firstname}
+                            {firstName}
                         </Typography>
                     </Box>
                     <Box m={1}>
                         <Typography>
-                            {lastname}
+                            {lastName}
                         </Typography>
                     </Box>
                     <Divider />
