+++ src/components/organization/OrganizationsPartnerList.js
@@ -1,19 +1,26 @@
 import {Controller, useForm} from "react-hook-form";
 import {Box, Button, FormControl, Grid, IconButton, InputLabel, MenuItem, Select, TextField} from "@material-ui/core";
 import {withSwal} from "../formsCommons/WithSwal";
-import React, {useEffect} from "react";
+import React, {useEffect, useState} from "react";
 import ClearIcon from "@material-ui/icons/Clear";
 import {OrganizationPartner} from "./OrganizationPartner";
 import {usersAPI} from "../../utils/apis/UserApi";
+import {validateEmail} from "../../utils/functions";
+import {PartnerForm} from "./PartnerForm";
 
 export const OrganizationsPartnerList = (props) => {
 
     const id = 1 //TODO change after jwt
-    function updateUsers(data) {
-        data.map((el => {
-            users.push({firstName: el.user.firstName, lastName: el.user.lastName, mail: el.user.mail})
-        }))
 
+    const [users, setUsers] = useState([props.users])
+    const partnerForm = PartnerForm({
+        users: users,
+        userAdded: (user) => setUsers(user)
+    })
+
+    function updateUsers(data) {
+        setUsers(data.map(el => el.user))
+        partnerForm.setUsers(users)
     }
 
     useEffect(() => {
@@ -24,42 +31,15 @@ export const OrganizationsPartnerList = (props) => {
         }
     }, [id])
 
-    const defaultFormState = {
-        firstName: "",
-        lastName: "",
-        mail: ""
-    }
-
-
-    const users = props.users || []
-    const {handleSubmit, control, reset} = useForm({defaultValues:defaultFormState})
-
-    const onUserAdd = (data) => {
-        withSwal({
-            loadingTitle: "Adding HR user",
-            promise: () => props.onSubmit(data),
-            successSwalTitle: "Success",
-            successSwalText: "You've successfully added HR user!",
-            successFunction: () => {
-                props.onChange(users.concat([data]))
-                reset()
-            },
-            errorSwalTitle: "We couldn't save this user for you"
-        })
-        if(props.onSubmit){
-            props.onSubmit(data)
-        }
-    }
 
     const onUserDelete = (idx, data) => {
         withSwal({
             loadingTitle: "Deleting HR user",
-            promise: () => props.onDelete(data),
+            promise: () => usersAPI.deleteHrPartner(data),
             successSwalTitle: "Success",
             successSwalText: "You've successfully deleted HR user!",
             successFunction: () => {
-                props.onChange(users.slice(0, idx).concat(users.slice(idx+1, undefined)))
-                reset()
+                setUsers(users.slice(0, idx).concat(users.slice(idx+1, undefined)))
             },
             errorSwalTitle: "We couldn't delete this user for you"
         })
@@ -69,65 +49,7 @@ export const OrganizationsPartnerList = (props) => {
     }
 
     return <>
-        <form id="user-form" name="user-form" onSubmit={handleSubmit(onUserAdd)}></form>
-        <Grid item xs={12} sm={6}>
-            <Controller
-                control={control}
-                name="firstName"
-                rules={{required: true}}
-                defaultValue={defaultFormState.firstName}
-                render={({field}) =>
-                    <TextField
-                        {...field}
-                        label="First Name"
-                        variant="outlined"
-                        form="user-form"
-                        fullWidth />
-                }
-            />
-        </Grid>
-        <Grid item xs={12} sm={6}>
-            <Controller
-                control={control}
-                name="lastName"
-                rules={{required: true}}
-                defaultValue={defaultFormState.lastName}
-                render={({field}) =>
-                    <TextField
-                        {...field}
-                        label="First Name"
-                        variant="outlined"
-                        form="user-form"
-                        fullWidth />
-                }
-            />
-        </Grid>
-        <Grid item xs={12} sm={6}>
-            <Controller
-                control={control}
-                name="mail"
-                rules={{required: true, validate: (mail) => !users.some(u => u.mail === mail)}}
-                defaultValue={defaultFormState.lastName}
-                render={({field}) =>
-                    <TextField
-                        {...field}
-                        label="e-mail address"
-                        variant="outlined"
-                        form="user-form"
-                        fullWidth />
-                }
-            />
-        </Grid>
-
-
-        <Grid item xs={12} sm={3}>
-            <Button
-                // className={classes.button}
-                type="submit"
-                variant="contained"
-                form="user-form"
-                fullWidth>Add HR user</Button>
-        </Grid>
+        <PartnerForm users={users} userAdded={(user) => setUsers(users.concat(user))}></PartnerForm>
         <Grid item xs={12}>
             <Box mt={0} mb={0}>
                 <Grid container>
@@ -135,7 +57,6 @@ export const OrganizationsPartnerList = (props) => {
                         <Grid key={idx} item xs={12} sm={6} md={4} lg={3}>
                             <Grid container>
                                 <Grid item xs={11}>
-                                    {/*<OfferSkill key={idx} name={skill.name} skillLevel={skill.level}/>*/}
                                     <OrganizationPartner key={idx} firstName={user.firstName} lastName={user.lastName} mail={user.mail}/>
                                 </Grid>
                                 <Grid item xs={1}>
