+++ src/utils/functions.js
@@ -8,6 +8,11 @@ export const scrollToTop = () => {
     document.documentElement.scrollTop = 0;
 }
 
+export const validateEmail = (email) => {
+    const re = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+    return re.test(String(email).toLowerCase())
+}
+
 export const getFilterValueByType = (filters, filterType) =>
     filters?.find(filter => filter.type === filterType)?.value
